Server - Updates Change Stream Functions
Isolate Change Stream functions to not import with other files
